Code refactoring, moving functionality to common location
