+++ test/unit/utils/caseTabTypeHelper.js
@@ -1,3 +1,6 @@
+const { uniqWith, filter, isEqual, map } = require('lodash');
+const { assert } = require('chai');
+
 const nonProdTabDisplayOrder = {
   History: 1,
   petitionDetails: 2,
@@ -118,9 +121,69 @@ const prodTabIds = [
   'ConfidentialDocuments'
 ];
 
+/**
+ * Validates the tab display ids to ensure there are no duplicate in the feature files
+ * @param tabIds List of case type TabID's
+ * @param caseTypeTab List of casetype tabs. A concatenation from all feature files
+ * @returns {[]} List with duplicates if found
+ */
+function validateUniqueTabDisplayOrder(tabIds, caseTypeTab) {
+  const errors = [];
+  let allFieldsPerTab = null;
+  let uniqResults = null;
+
+  tabIds.forEach(tabId => {
+    allFieldsPerTab = filter(caseTypeTab, field => {
+      return field.TabID === tabId;
+    });
+
+    uniqResults = uniqWith(
+        allFieldsPerTab,
+        (field1, field2) => {
+          return field1.TabFieldDisplayOrder === field2.TabFieldDisplayOrder;
+        });
+
+    if (!isEqual(allFieldsPerTab, uniqResults)) {
+      errors.push(uniqResults);
+    }
+  });
+
+  return errors;
+}
+
+/**
+ * Validates the tab field display order to ensure there are no gaps in feature files.
+ * Would throw an error if a duplicate tab display id is found
+ * @param tabIds List of case type TabID's
+ * @param caseTypeTab caseTypeTab List of casetype tabs. A concatenation from all feature files
+ */
+function validateTabFieldDisplayOrder(tabIds, caseTypeTab) {
+  let allFieldsPerTab = [];
+  let allTabFieldDisplayOrderNumbers = [];
+
+  tabIds.forEach(tabId => {
+    allFieldsPerTab = filter(caseTypeTab, field => {
+      return field.TabID === tabId;
+    });
+    allTabFieldDisplayOrderNumbers = map(allFieldsPerTab, field => {
+      return field.TabFieldDisplayOrder;
+    })
+        .sort((a, b) => {
+          return a - b;
+        });
+    for (let i = 1; i < allTabFieldDisplayOrderNumbers.length; i++) {
+      if (allTabFieldDisplayOrderNumbers[i] - allTabFieldDisplayOrderNumbers[i - 1] !== 1) {
+        assert.fail(`Missing/unordered TabFieldDisplayOrder sequence number in TabID ${tabId} - expected ${allTabFieldDisplayOrderNumbers[i - 1] + 1} but got ${allTabFieldDisplayOrderNumbers[i]}`);
+      }
+    }
+  });
+}
+
 module.exports = {
   nonProdTabDisplayOrder,
   nonProdTabIds,
   prodTabDisplayOrder,
-  prodTabIds
+  prodTabIds,
+  validateUniqueTabDisplayOrder,
+  validateTabFieldDisplayOrder
 };
